+++ ToDo/Controller/CategoryViewController.swift
@@ -7,13 +7,11 @@
 
 import UIKit
 import RealmSwift
-import SwipeCellKit
 
-class CategoryViewController: UITableViewController {
+class CategoryViewController: SwipeTableViewController {
     
     let realm = try! Realm()
     var tasksCategories: Results<CategoryOfTasks>?
-    
     
     override func viewDidLoad() {
         super.viewDidLoad()
@@ -32,9 +30,20 @@ class CategoryViewController: UITableViewController {
     }
     
     private func loadCategory() {
-        
         tasksCategories = realm.objects(CategoryOfTasks.self)
         tableView.reloadData()
+    }
+    
+    override func updateModel(at indexPath: IndexPath) {
+        if let categoryOFDeletion = self.tasksCategories?[indexPath.row] {
+            do {
+                try self.realm.write {
+                    self.realm.delete(categoryOFDeletion)
+                }
+            } catch {
+                print(error.localizedDescription)
+            }
+        }
     }
     
     @IBAction func addButtonPressed(_ sender: UIBarButtonItem) {
@@ -52,12 +61,10 @@ class CategoryViewController: UITableViewController {
             newTaskCategory.name = textField.text ?? ""
             self.save(category: newTaskCategory)
         }
-        
         alert.addTextField { alertTextField in
             alertTextField.placeholder = "Please create a new task group"
             textField = alertTextField
         }
-        
         alert.addAction(action)
         present(alert, animated: true)
     }
@@ -71,10 +78,8 @@ extension CategoryViewController {
     }
     
     override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        guard let cell = tableView.dequeueReusableCell(withIdentifier: "CategoryCell", for: indexPath) as? SwipeTableViewCell else { return UITableViewCell()}
-        
+        let cell = super.tableView(tableView, cellForRowAt: indexPath)
         cell.textLabel?.text = tasksCategories?[indexPath.row].name ?? "No categories added"
-        cell.delegate = self
         return cell
     }
 }
@@ -90,32 +95,5 @@ extension CategoryViewController {
     
     override func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
         performSegue(withIdentifier: "goToTasks", sender: self)
-    }
-}
-
-// MARK: - Swipe Table View Cell Delegate
-extension CategoryViewController: SwipeTableViewCellDelegate {
-    func tableView(_ tableView: UITableView, editActionsForRowAt indexPath: IndexPath, for orientation: SwipeActionsOrientation) -> [SwipeAction]? {
-        guard orientation == .right else { return nil }
-        
-        let deleteAction = SwipeAction(style: .destructive, title: "Delete") { action, indexPath in
-            if let categoryOFDeletion = self.tasksCategories?[indexPath.row] {
-                do {
-                    try self.realm.write {
-                        self.realm.delete(categoryOFDeletion)
-                    }
-                } catch {
-                    print(error.localizedDescription)
-                }
-            }
-        }
-        deleteAction.image = UIImage(named: "delete")
-        return [deleteAction]
-    }
-    
-    func tableView(_ tableView: UITableView, editActionsOptionsForRowAt indexPath: IndexPath, for orientation: SwipeActionsOrientation) -> SwipeOptions {
-        var options = SwipeOptions()
-        options.expansionStyle = .destructive
-        return options
     }
 }
